refactor: 잘못된 경로 재설정
