+++ src/app/(post)/list/[id]/(components)/PostDetail.tsx
