+++ src/utils/postApi.ts
@@ -1,8 +1,8 @@
 import { ImageType, WriteTypes } from "@/types/write";
-import supabase from "./supabase";
+import browserClient from "./supabase/client";
 
 export const getPost = async (id: string): Promise<WriteTypes | null> => {
-  const { data, error } = await supabase.from("post").select("*").eq("post_id", id).single();
+  const { data, error } = await browserClient.from("post").select("*").eq("post_id", id).single();
 
   if (error) {
     console.error("데이터 요청 오류:", error);
@@ -18,7 +18,7 @@ export const uploadImage = async (imageState: ImageType) => {
     throw new Error("이미지 파일이 없습니다.");
   }
 
-  const { data, error } = await supabase.storage
+  const { data, error } = await browserClient.storage
     .from("post")
     .upload(`post_${imageState.imageFile.lastModified}`, imageState.imageFile);
 
@@ -28,7 +28,7 @@ export const uploadImage = async (imageState: ImageType) => {
 
 // 게시글 추가 함수
 export const addPost = async (formData: Omit<WriteTypes, "fileInputRef">) => {
-  const { data, error } = await supabase.from("post").insert(formData).select("*");
+  const { data, error } = await browserClient.from("post").insert(formData).select("*");
 
   if (error) throw error;
   return data;
